Oracle driver

+++ 4.Samples/Vita.Samples.BookStore/BooksModule/Api/Models/Models.cs
@@ -71,7 +71,7 @@ namespace Vita.Samples.BookStore.Api {
     public Guid Id;
     public Guid OrderId; 
     public Book Book;
-    public byte Quantity; // byte? - see comment for IBookOrderLine.Quantity property
+    public int Quantity; 
   }
 
   [DebuggerDisplay("{UserName}")]
